requirements.txt
python-twitter==2.2
selenium==2.52.0